Add setup script and configuration/service file templates
This commit adds a network configuration template along with an SSH
tunneling .service file template and a .service file for the scanner
script.

Additionally, a setup script is provided which will prompt for some
information before setting up the configuration based on the templates.
If the configuration templates change, the setup script can be run again
to update the configuration.