Use a request queue thread.
As previously designed, the script would start a thread for every read.
Aside from being wasteful, these threads would all individually restart
the wifi connection in case they failed, causing multiple failing
check-ins to generate a barrage of wifi connects and disconnects.
Additionally, in some cases where a user performs a check-in and
check-out sequence in rapid succession when the wifi connection is being
restarted, the check-in could be processed before the check-out, causing
the system to enter an invalid state.

The new design uses a request deque, which is processed in a single
separate thread. Failing requests are added back to the left end of the
deque, and new requests are added to the right end. This way, requests
can not be processed out-of-order. In addition, when a request fails,
the wifi connection is restarted a single time, and the request queue
thread waits for this process to finish before processing other
requests.

Moreover, time-outs are added to the wifi connection restart
process, so that the script cannot hang indefinitely there. If the
restart fails, it will simply be retried when the previously failing
request is retried and fails again.

Lastly, the new design uses Python's type annotations, so the code can
be statically analyzed by tools such as mypy.