read_events: require a config file (config.json), and add an example (config.json.example).
One of the long-standing issues has been dealing with restarting Pi's and the order in which scanners were detected.
This commit resolves that issue by requiring us to declare the scanners as in- or output by ID, instead of simply using event0 and event1.
Additionally, this moves the definition of the endpoint and zone into a config file, instead of into the source.